requirements.txt
absl-py==0.9.0
appdirs==1.4.3
apturl==0.5.2
asttokens==2.4.1
astunparse==1.6.2
atomicwrites==1.1.5
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bleach==3.1.1
blinker==1.4
blosc==1.7.0
Brlapi==0.7.0
cachetools==4.0.0
caffe==1.0.0
catkin-pkg==0.4.16
certifi==2024.2.2
cffi==1.14.0
chardet==3.0.4
Click==7.0
cloudpickle==1.3.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
dask==2.8.1+dfsg
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
dill==0.3.1.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
docopt==0.6.2
docutils==0.16
duplicity==0.8.12.0
empy==3.3.2
entrypoints==0.3
et-xmlfile==1.0.1
exceptiongroup==1.0.4
executing==2.0.1
fasteners==0.14.1
fastjsonschema==2.19.1
filelock==3.0.12
flake8==3.7.9
fsspec==0.6.1
future==0.18.2
gast==0.3.3
gdown==3.12.2
gencpp==0.6.2
genlisp==0.4.17
genmsg==0.5.14
genpy==0.6.9
google-auth==1.5.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.1
h5py==2.10.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
imageio==2.4.1
importlib_metadata==7.1.0
iniconfig==1.1.1
ipykernel==5.2.0
ipython==8.12.3
ipython_genutils==0.2.0
ipywidgets==6.0.0
jdcal==1.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jsonschema==3.2.0
jupyter-console==6.0.0
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kaggle==1.6.11
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
locket==0.2.0
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.1.5
matplotlib==3.1.2
matplotlib-inline==0.1.7
mccabe==0.6.1
mistune==3.0.2
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
netifaces==0.10.4
networkx==2.4
nose==1.3.7
notebook==6.0.3
numexpr==2.7.1
numpy==1.24.4
nvidia-ml-py==7.352.0
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.3
opt-einsum==3.2.1
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
parameterized==0.7.0
paramiko==2.6.0
parso==0.8.4
partd==1.0.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
pipreqs==0.5.0
platformdirs==4.2.1
pluggy==0.13.0
ply==3.11
POT==0.4.0
prometheus-client==0.7.1
prompt-toolkit==3.0.43
protobuf==3.11.4
psutil==5.5.1
pure-eval==0.2.2
py==1.8.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.19
pycups==1.9.73
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.17.2
PyGObject==3.36.0
pygpu==0.7.6
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyRFC3339==1.1
pyrsistent==0.15.5
PySocks==1.7.1
pytest==7.2.0
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.9.0.post0
python-debian===0.1.36ubuntu1
python-slugify==8.0.4
pytz==2019.3
PyWavelets==0.5.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==26.0.3
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.0.0
requests-unixsocket==0.2.0
roman==2.0.0
rospkg==1.2.3
rsa==4.0
scikit-image==0.16.2
scikit-learn==1.3.2
scipy==1.10.1
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
ssh-import-id==5.10
stack-data==0.6.3
systemd-python==234
tables==3.6.1
tblib==1.6.0
tensorboard==2.3.0
tensorflow-estimator==2.3.0
tensorflow-gpu==2.3.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.4
text-unidecode==1.3
Theano==1.0.4
threadpoolctl==3.4.0
tinycss2==1.3.0
tomli==2.0.1
toolz==0.9.0
torch==1.6.0
torchvision==0.7.0
tornado==6.4
tqdm==4.59.0
traitlets==5.14.3
typing_extensions==4.11.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==2.0.0
wrapt==1.11.2
xkit==0.0.0
xlrd==1.1.0
xlwt==1.3.0
yarg==0.1.9
zipp==1.0.0
